+++ packages/hashmap/bsconfig.json
@@ -1,0 +1,21 @@
+{
+  "name": "rescript-hashmap",
+  "sources": [
+    {
+      "dir": "src",
+      "subdirs": true
+    },
+    {
+      "dir": "tests",
+      "type": "dev"
+    }
+  ],
+  "package-specs": [
+    {
+      "module": "es6",
+      "in-source": true
+    }
+  ],
+  "suffix": ".mjs",
+  "external-stdlib": "@rescript/std"
+}
