+++ packages/hashmap/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "rescript-hashmap",
+  "version": "0.1.0",
+  "description": "Persistent Hash-map implemenetation in ReScript",
+  "license": "MIT",
+  "author": {
+    "name": "Hyunwoo Nam"
+  },
+  "keywords": [
+    "rescript",
+    "data-structure",
+    "immutable",
+    "hashmap"
+  ],
+  "scripts": {
+    "postinstall": "rescript build -with-deps",
+    "build": "rescript build",
+    "prepack": "yarn build && yarn bundle",
+    "start": "rescript build -w",
+    "clean": "rescript clean"
+  },
+  "files": [
+    "dist",
+    "src/*.res",
+    "src/*.resi",
+    "bsconfig.json"
+  ],
+  "dependencies": {
+    "@rescript/std": "^9.1.3"
+  },
+  "devDependencies": {
+    "rescript": "^9.1.4"
+  }
+}
