+++ packages/hashmap/src/HashMap.mjs
