+++ packages/hashmap/src/HashMap.resi
