+++ packages/hashmap/src/HashSet.mjs
@@ -1,16 +1,16 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as $$Map from "./Map.mjs";
+import * as HashMap from "./HashMap.mjs";
 import * as Caml_option from "@rescript/std/lib/es6/caml_option.js";
 
 function make(hasher) {
   return {
-          hashMap: $$Map.make(hasher)
+          hashMap: HashMap.make(hasher)
         };
 }
 
 function get(s, v) {
-  var match = $$Map.get(s.hashMap, v);
+  var match = HashMap.get(s.hashMap, v);
   if (match !== undefined) {
     return Caml_option.some(v);
   }
@@ -19,18 +19,18 @@ function get(s, v) {
 
 function set(s, v) {
   return {
-          hashMap: $$Map.set(s.hashMap, v, undefined)
+          hashMap: HashMap.set(s.hashMap, v, undefined)
         };
 }
 
 function remove(s, v) {
   return {
-          hashMap: $$Map.remove(s.hashMap, v)
+          hashMap: HashMap.remove(s.hashMap, v)
         };
 }
 
 function size(s) {
-  return $$Map.size(s.hashMap);
+  return HashMap.size(s.hashMap);
 }
 
 export {
