+++ packages/hashmap/src/impl/Hamt.mjs
