+++ packages/hashmap/tests/hamt_test.mjs
@@ -1,6 +1,6 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Hamt from "./Hamt.mjs";
+import * as Hamt from "../src/impl/Hamt.mjs";
 import * as Caml_obj from "@rescript/std/lib/es6/caml_obj.js";
 import * as Belt_List from "@rescript/std/lib/es6/belt_List.js";
 import * as Belt_Array from "@rescript/std/lib/es6/belt_Array.js";
