+++ packages/hashmap/tests/hamt_test.res
